Refactor apollo provider to use auth0
