+++ src/components/PrivateRoute.js
@@ -1,15 +1,6 @@
-import React, { useState, useEffect } from 'react';
-import gql from 'graphql-tag';
-import localforage from 'localforage';
-import { useQuery } from '@apollo/react-hooks';
+import React, { useEffect } from 'react';
 import { useAuth0 } from '../react-auth0-spa';
 import { Route, Redirect } from 'react-router-dom';
-
-const GET_AUTH = gql`
-  {
-    loggedIn @client
-  }
-`;
 
 const PrivateRoute = ({ component: Component, path, ...rest }) => {
   const { loading, isAuthenticated, loginWithRedirect } = useAuth0();
